just make the variable signed.
svn:sync-xref-src-repo-rev: r22689

+++ src/VBox/Devices/PC/DevRTC.cpp
@@ -1,4 +1,4 @@
-/* $Id: DevRTC.cpp 3314 2007-06-27 11:44:57Z noreply@oracle.com $ */
+/* $Id: DevRTC.cpp 3477 2007-07-05 15:18:17Z knut.osmundsen@oracle.com $ */
 /** @file
  * Motorola MC146818 RTC/CMOS Device.
  */
@@ -141,7 +141,7 @@ struct RTCState {
     /** Number of release log entries. Used to prevent flooding. */
     uint32_t        cRelLogEntries;
     /** The current/previous timer period. Used to prevent flooding changes. */
-    uint32_t        CurPeriod;
+    int32_t         CurPeriod;
 };
 
 #ifndef VBOX_DEVICE_STRUCT_TESTCASE
@@ -169,7 +169,7 @@ static void rtc_timer_update(RTCState *s, int64_t current_time)
         s->next_periodic_time = ASMMultU64ByU32DivByU32(next_irq_clock, freq, 32768) + 1;
         TMTimerSet(s->CTXSUFF(pPeriodicTimer), s->next_periodic_time);
 
-        if ((uint32_t)period != s->CurPeriod)
+        if (period != s->CurPeriod)
         {
             if (s->cRelLogEntries++ < 64)
                 LogRel(("RTC: period=%#x (%d) %u Hz\n", period, period, _32K / period));
